backend/requirements.txt
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.5
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.21.6
packaging==23.2
pymongo==4.6.1
python-decouple==3.8
python-dotenv==0.21.1
scipy==1.7.3
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
